# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Prometheus2 Backend API
  version: 0.2.0
  description: API backend pura JSON/CSV per Prometheus2.
servers:
  - url: http://localhost:8080
paths:
...
  /controls:
    get:
      summary: Lista controlli paginata
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: per_page, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: sort, in: query, schema: { type: string } }
        - { name: direction, in: query, schema: { type: string, enum: [asc, desc] } }
        - { name: registry_number, in: query, schema: { type: string } }
        - { name: registry_year, in: query, schema: { type: string } }
        - { name: date_from, in: query, schema: { type: string, format: date } }
        - { name: date_to, in: query, schema: { type: string, format: date } }
        - { name: outcome, in: query, schema: { type: string, enum: [positivo, negativo, in_accertamento] } }
        - { name: status, in: query, schema: { type: string, enum: [bozza, validato, annullato] } }
      responses:
        "200":
          description: Lista paginata controlli
    post:
      summary: Crea controllo in bozza
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ControlCreateRequest"
      responses:
        "201":
          description: Controllo creato
        "422":
          description: Validazione non riuscita
...
  /controls/{control}/validate:
    post:
      summary: Valida controllo
      parameters:
        - name: control
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Controllo validato
  /controls/{control}/annul:
    post:
      summary: Annulla logicamente controllo
      parameters:
        - name: control
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Controllo annullato
...
components:
  schemas:
...
    ControlCreateRequest:
      type: object
      required: [_csrf_token, control_date, control_time, has_event, business_name, business_location, primary_category_id, outcome]
      properties:
        _csrf_token: { type: string }
        control_date: { type: string, format: date }
        control_time: { type: string, example: "10:30" }
        has_event: { type: integer, enum: [0, 1] }
        event_name: { type: string }
        business_name: { type: string }
        business_location: { type: string }
        primary_category_id: { type: integer }
        secondary_category_ids:
          type: array
          items: { type: integer }
        agent_ids:
          type: array
          items: { type: integer }
        outcome: { type: string, enum: [positivo, negativo, in_accertamento] }